orchestrator/requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-amazon==8.27.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-cncf-kubernetes==8.4.1
boto3==1.34.131
google-api-python-client==2.147.0
dvc==3.51.2
langsmith==0.1.77
prometheus-client==0.20.0
